requirements.txt
streamlit>=1.31
sqlalchemy>=2.0
psycopg2-binary   # или asyncpg, если используете async
pgvector
python-dotenv
google-api-python-client
openai
scikit-learn
numpy